move TestCLIGenerateEquivalence to ut
